changed data type from uint to float

+++ GPUSorting/Shaders/ComputeShader2.hlsl
@@ -1,5 +1,5 @@
-RWStructuredBuffer<uint> outputData : register(u0, space0);
-StructuredBuffer<uint> srv : register(t0, space0);
+RWStructuredBuffer<float> outputData : register(u0, space0);
+StructuredBuffer<float> srv : register(t0, space0);
 [numthreads(1, 1, 1)]
 void CS_main( uint3 DTid : SV_DispatchThreadID )
 {
